Moved Marker rendering into separate view
